// src/app/Interfaces/Request/addClient.interface.ts
export interface addClient {
  id: null;
  registrationDate: string;
  cancellationDate: string;
  serviceCharge: number;
  depositAccount: string;
  chargeAccount: string;
  monthlyFee: number;
  email: string;
  sendFile: number;
  sendMonthlyFeeFile: number;
  active: number;
  transactionsLimit: string;
  deliveryTime: string;
}